Import from CVS Tue Dec 14 10:22:27 2004 -0600
- Added extra info and dist info for binary rpms from common source rpm.
file .rpmdist.in was initially added on branch LIS-2-16-18-autoconf.
file .rpmextra.in was initially added on branch LIS-2-16-18-autoconf.
file .rpmdist.in was initially added on branch HP.
file .rpmextra.in was initially added on branch HP.
file .rpmdist.in was initially added on branch OpenSS7-0_9_2.
file .rpmextra.in was initially added on branch OpenSS7-0_9_2.